Recompose components and dislay saved pictures in new page
